+++ src/main/java/org/comit/pluralisticsecurity/PluralisticSecurityApplication.java
@@ -1,60 +1,12 @@
 package org.comit.pluralisticsecurity;
 
-import java.util.ArrayList;
-
-import org.comit.pluralisticsecurity.entity.Role;
-import org.comit.pluralisticsecurity.entity.RoleEnum;
-import org.comit.pluralisticsecurity.entity.User;
-import org.comit.pluralisticsecurity.entity.UserRole;
-import org.comit.pluralisticsecurity.repository.UserRepository;
-import org.comit.pluralisticsecurity.repository.UserRoleRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 @SpringBootApplication
-public class PluralisticSecurityApplication implements CommandLineRunner {
-
-	@Autowired
-	private UserRepository userRepository;
-
-	@Autowired
-	private UserRoleRepository userRoleRepository;
+public class PluralisticSecurityApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(PluralisticSecurityApplication.class, args);
 	}
-
-	public void run(String... args) {
-		UserRole userRole = new UserRole();
-		UserRole adminAccount = userRoleRepository.findByRole(Integer.valueOf(RoleEnum.ADMIN.ordinal()));
-		//UserRole adminAccount = userRoleRepository.findByRole(RoleEnum.ADMIN.toString());
-
-		if (null == adminAccount) {
-
-			User user = new User();
-			Role role = new Role(RoleEnum.ADMIN.toString());
-			//Role role = new Role(RoleEnum.ADMIN.ordinal());
-
-			user.setEmail("admin@gmail.com");
-			user.setFirstname("admin");
-			user.setLastname("admin");
-			user.setUsername("admin");
-			user.setActive(true);
-
-			user.setPassword(new BCryptPasswordEncoder().encode("admin"));
-			userRole.setRole(role);
-			userRole.setUser(user);
-
-			user.setUserRoles(new ArrayList<>());
-			user.getUserRoles().add(userRole);
-
-			userRepository.save(user);
-
-		}
-
-	}
-
 }
